console: Move webhook format logic middleware
